Make UI compatible with remote servers, added in powa4
As a consequence, bump version of powa-web to 4.0.0 (for now in beta).

A new screen is added as the first screen to choose the server to display.

All datasources now use an additional qual "srvid = ?" to get the data related
to the chosen server.

Also teach the queries that needs to connect on the real database (qual
resolution, explain, wizard...) to do so.

The configuration page get its own hierarchy.  The first level now displays
general information about the overall setup (status of powa-collector and the
background worker, list of remote servers with their status and any errors
reported by powa-collector and the snapshot functions), and each server get its
own configuration page to display the underlying extensions and PostgreSQL
configuration.

The datasources can now also reports error messages (in the returned
data["alerts"]), that will be asynchrounously displayed on the current page.
This is for now only used in the root configuration page, to warn if no
collector is running.

While at it, also do some minor changes to the general design, to remove some
wasted space and use foundation icons and CSS coloring for boolean field
representation instead of UTF8 characters.

Many thanks to Adrien Nayrat for thorough testing.