Query and display powa-collector information
powa-collector has a communication protocol to ask for various information, so
add a dedicated notify_collector function to query it through this protocol and
retrieve the result as an array of answers.  It uses a random channel for each
call to make sure that we won't get answers from concurrent queries.

Use this infrastructure to display powa-collector state in various /config
pages.

Most of the powa-collector related functions are in a new collector.py file.