Avoid connecting to remote server when possible for checking extension
Rename has_extension to has_extension_version, which connect to the given,
possibly remote, server, on the given database, and returns the installed
version if any

Also add a new has_extension function, which doesn't accept a database
parameter, which will always connect on the dedicated powa server (either local
or repository), and will check if the extension appears to be installed on the
remote server by looking at powa_functions configuration for the given remote
server.  As we don't connect to the remote servers, we assume that the
powa_functions is correctly setup.  Event if that's not the case, it's not
really a problem as the UI query will have the underlying objects to query will
always exists on the dedicated powa repo database (except for
pg_track_settings, but the code already check for local extension availability
for this reason, instead of the remote one).