Handle connection for user objects authorization.
This adds a new "remote_conn" parameter in execute(), connect() (and a few
other functions) to let caller know if the connection is used to check users
objects or not.  This is check against a new "allow_ui_connection" global
setting in the configuration file to globally disable such connections, or the
recently introduced per-server "allow_ui_connection" configuration in
powa-archivist 4.0.0.

If the connection is denied, fail with an appropriate message to let UI users
know why some widget are not working.