Always tag connections to non local server as remote access.
This makes sure that we properly forbid most of remote access, even if the
caller forget to specify the remote_access flag, as this is already the case
for at least get_pg_version().

The parameter is still needed as the access to local server but on a database
different than the dedicated one is still a remote access, that should be
properly handled.