Fix per-query hypothetical index checking.
Thanks to github user hrawulwa.

This part wasn't providing the srvid, thus possibly trying to call
pg_qualstats_example_query() on the wrong server.  Also, as this can now be
called on a remote server, we can't rely on having powa_statements table
filled, so change the query to only call pg_qualstats_example_query().  While
at it add some safety checks to make sure that we correctly handle cases when
no example query can be found, so we still display the information about
possible index to be created.