Add compatibility with pg_stat_statements 1.8
Also expose all new counters.

Also rely on the extension version stored in PoWA tables since v4.1.0 to avoid
retrieving the version from remote servers.  This avoids slowdown in the UI and
also brings more feature when the UI isn't allowed to connect on the remote
servers.

The has_extension_version is changed to accept a new parameter for the minimum
version, rather than doing the test in each call site.