Use locally available info in config.py when possible.
If powa >= 4.1, try first to get data from powa_extension table to get extension
status rather than trying to connect on the remote server.  Note that it's
still possible that powa-collector hasn't been updated yet, leading to globally
unknown status.  Config.py isn't a critical part of the UI, so don't try to
have a better handling of that situation for now.

If pg_track_settings >= 2.0 is installed, try to get the GUC values locally
rather than connecting on the remote server.  Note that the GUC category and
unit are retrieved from the repository server, so some discrepancy is possible
if the repository server and the remote servers are from different PostgreSQL
versions.

Those changes should improve latency, and give information on setup where
remote connction is not possible.