Allow missing pg_stat_kcache records.
The UI will always try to join records from pg_stat_statements and
pg_stat_kcache when pg_stat_kcache is enabled, but it's possible that
records from pg_stat_kcache are missing.  One current scenario is when showing
information about a utility statements, are those are for now not handled by
pg_stat_kcache.

Ideally pg_stat_kcache should be fixed to handle those, but in the meantime
let's just support such case by change the JOIN to a LEFT OUTER JOIN.

Thanks to Christoph Dreis for the report.