Fix qualstat_get_figures() when called with int queryids.
The call to qualstat_get_figures() in sql/__init__.py could call that function
with raw values from powa_statements, and the function was originally written
to be fed with string values, in powa/query.py.

This is a long standing bug, which has been unnotice for quite some time as
this code path is only used if no example query is found with pg_qualstats.

Thanks to Martin Aparicio and Frédéric Yhuel for the bug report.