qualify extension's relation part 2: get rid of sqlalchemy query generator.
With the need to properly handle schema-qualification, it's just easier to
fully work with SQL rather than a mix of SQL and sqlalchemy-specific code.

With this commit all the code is now properly schema qualifying object.

The only remaining usage of sqlalchemy is for parameters, so a following commit
will entirely remove sqlalchemy and rely on psycopg.  As a preliminary step,
all the dynamic code is already using the format() function rather than using
old style %()s formatting, as psycopg named parameters rely on that format.