wizard: correctly handle multiple query with the same qual
When building the initial list of quals we add quals with the qualid as the
collection id, which is a unique identifier.

When another qual with the same qualid is found, we need to merge it to the
existing qual in the collection, if any, so we can later check the related
query with the found hypothetical indexes.