Remove sqlalchemy dependency.
Parameters are now entirely handled with psycopg using their syntax (%(prm)s
rather than :prm), so there's no need for sqlalchemy anymore.

As a consequence, we now want the regular "user" parameter, rather than
"username".  We already accepted both, so invert the workaround so that "user"
is the default name and "username" the accepted alias.

Also, MetricGroupDef's query property can now either return a plain query
string or a (query, parameters) tuple if specific parameters needs to be
associated with the query.

To simplify a bit parameters handling, a default "samples=100" parameter is
added rather than specifying it everywhere in the code.  This will also make it
easier to handle custom samples size if we ever want to do that.

Finally, the execute() function will now return a list of RealDictRow.