Fix detection of powa bgworker.
If a regular backend calls the powa_take_snapshot(), the application_name will
automatially be updated, leading to that backend be considered as the bgworker,
or one of multiple bgworker.  So make sure that only backend type "powa" (which
is set by the bgworker) is used for that purpose.

Detection of whether a collector is running for a specific server is still
broken, as this part has never been updated for remote servers /
powa-collector, but I will fix that in a later commit.