Improve grid queries performance.
To improve performance, the retrieval of the coalesced records is divided
in 3 parts, which can greatly limit the number of rows that will need to be
sorted.

Those 3 parts are split in 3 subqueries.  For the left and right bounds
(the smallest and largest timestamp in the given interval), we get the 2
underlying records in the coalesced records (if any) and unnest them
fully, filtering any records outside of the interval bound.  For any record
in the coalesced records that are entirely inside the given interval,
we rely on the pre-computed metadata (mins_in_range and maxs_in_range) and
simply return that, which can greatly reduces the amount of rows to sort.

For the "current" records, we simply return all the rows in the given
interval.

Marc Cousin and Julien Rouhaud.