Add dark/light theme switch
dark_theme is stored in the localStorage to keep styling during
the navigation.
We have removed v-main classes in layout.html to manage the dark
and light background on the v-application element.
We copy the stackoverflow highlight dark and light css into a new
highlight.scss file to dynamically manage highlight.css.
We will be able to dynamically import the source file with
SASS 1.64, but this will require Vuetify 3.