Split views.py in multiple files
That files is getting quite big, the naming convention is a bit confusing, and
many new widgets are about to be added so it's time to cleanup things a bit.

Extract the parts related to the grid components into a new
powa/sql/views_grid.py file, same for the parts related to graphs in
powa/sql/views_graph.py file.

The naming convention isn't changed, but hopefully this separation will help to
make things easier to follow.