Add buttons to force a snapshot or register a catalog refresh from the UI
The new buttons, and the existing "reload collector", now have proper
permission checks.  This relies on the "powa_roles" feature introduced in
powa-archivist v5.  The roles that can use any of the buttons are:

- powa_signal_backend
- powa_admin
- powa_write_all_data

Those roles are however optional and may not be configured, so just in case for
pg14+ pg_signal_backend is also allowed.

The button isn't displayed if you're not have USAGE privilege on any of those
roles.  Note that powa-collector will also make sure that the requesting role
has enough permission, so hiding the button for the unauthorized role is just
meant to save a LISTEN/NOTIFY roundtrip to the collector, knowing that the only
response would be "permission denied".

Database names are user provided, so notify_collector now ensures that all the
parameters are properly escaped.