Add MetricGroupDef.split()
This method splits the metrics according to a list of list of metrics.  This is
useful when a MetricGroupDef is used for multiple widgets, where all the widget
use a disjoint set of metrics and one (or multiple) widgets just need to
contain a small subset while one of the widget needs everything else.

An upcoming commit adding support for pg_stat_activity graphs will benefit from
such a method.