Get rid of vue components in html templates
Everything now goes in .vue files.

We don't need to manually declare the Vuetify components anymore.

We now expose handler properties as JSON in template.

We can't use VuetifyResolver for treeshaking because of the
external component V-Snackbars. But we actually have our own resolver
which does the same except that it excludes VSnackbars.