Avoid reloading page, turn into single page app
By using vue-router (in HTML5 mode) we can reload the page content
without reloading the page itself.

The urls are still compatible with the old mechanism, which means that
an url copied from the url toolbar will still work and display the same
content.

Disabled attribute for breadcrumb items is not required anymore since
vuetify in combination with vue-router disables the item for the current
route.