Cache extension versions
More and more widgets need to adapt their behavior based on the remote server's
extension version (or availability).  While most of the time getting the
information only requires accessing local repository, it's not guaranteed and
is a bit of a waste of cycles anyway so cache the information when we retrieve
it.