Make sure toprow headers count match columns for grid
Previously we were expecting the first column not to be provided and to
be taken into account in the UI. We now explicitely set it server side.

Also get rid of merge property for toprow headers. It was relevant when we were
using backgrid but we don't use it anymore.