Fix wrong positioning of snackbars
Before this change, there's a lot space between snackbars.

This change is made to counterbalance a change in the teleport
positioning computation (ie. with the use of --v-layout-[bottom|top] for
the padding computation). We actually don't want Vuetify to computed the
positioning since we're using flex display on the parents to stack the
snackbars.