Fix version error message when building package
After the commit ecb8013b5f in which we changed the way the python code
is formatted, the "hack" to get the version in setup.py no longer
worked.

The current commit fixes this. A better management of the version should
be implemented. For example, using setuptools_scm could be a good
option.