Fix broken management of alert messages
In aded59171451ac, the function to add alert message has been modified
for no good reason resulting in message to be empty in the interface.

We also fix alert message management when loading data. This has been
missed when switching to VueJS. We shouldn't use "this" keyword anymore in
vue components or views.