fix: Fix notifiation authorization check on pg13-
Postgres 13 or lower don't have predefined roles, so we call unnest on an empty
python array which is emitted as '{}'.  This is ambiguous for postgres so we
need to explicitly cast the value as an array of text, which works for both an
empty array and a non empty array (which is emitted as ARRAY[...]).