Fetch data source only when component is mounted
This way, data for components that are hidden at first (for example in
inactive tabs) won't be fetched until required. We then save a bunch of
requests to the server, when navigating from one page to another,
without opening tabs.

On date range refresh, however, the data sources will be requested
again even if the component using the data is not visible anymore.