Improve data fetch (#261)
* Remove unrequired templates for loading progress bars

* Don't provide dashboardConfig

It is never injected.

* Improve data fetching

The main idea is to rely on a single store to manage data (dashboard,
data sources, and changes). It handles the data using Fetch API and is
able to abort requests. Navigating in the application or changing the
date range may abort pending requests.

This fixes a bug happening when navigating quickly in the application
(using the history back button in the browser) with components
(essentially grids) displaying the data for other pages.

We now use pinia which has a very nice API, and is commonly used in
VueJS projects.

In some components (for example Graph), we rely more on reactivity (ie.
computed values).

Along the way, we also add a loading bar for the dashboard (at the top
of the page).

* Fetch data source only when component is mounted

This way, data for components that are hidden at first (for example in
inactive tabs) won't be fetched until required. We then save a bunch of
requests to the server, when navigating from one page to another,
without opening tabs.

On date range refresh, however, the data sources will be requested
again even if the component using the data is not visible anymore.