Properly detect missing extensions
If the powa-extension is not installed on the repository server, login should
fail and the UI should report an error message.  Unfortunately the existing
logic was broken as the underlying function was throwing an exception rather
than returning an expected None if the powa extension is not installed.  This
commit fixes this case and properly return None.

While at it, report a better error if one of the needed extension is not
installed on the repository server.  An internal error stack is still
displayed, but at least it will show an informative "extenion not found" with
the associated extension name rather than an obscure "key error" message.